<==== Mercurial data ====>
Branch: mobile_client_1

+++ common/src/plugins/resource/onvif/onvif_resource.cpp
@@ -2688,7 +2688,8 @@ void QnPlOnvifResource::stopInputPortMonitoringAsync()
         asyncPullMessagesCallWrapper->join();
     }
 
-    QnSoapServer::instance()->getService()->removeResourceRegistration( toSharedPointer().staticCast<QnPlOnvifResource>() );
+    if (QnSoapServer::instance() && QnSoapServer::instance()->getService())
+        QnSoapServer::instance()->getService()->removeResourceRegistration( toSharedPointer().staticCast<QnPlOnvifResource>() );
 }
